+++ src/gafaelfawr/models/token.py
@@ -124,7 +124,9 @@ class TokenGroup(BaseModel):
         regex=GROUPNAME_REGEX,
     )
 
-    id: int = Field(..., title="The numeric GID of the group", example=123181)
+    id: Optional[int] = Field(
+        None, title="The numeric GID of the group", example=123181
+    )
 
 
 class TokenBase(BaseModel):